@@ -6,21 +6,21 @@ import 'container.dart';
 import 'text.dart';
 
 abstract class ElementLayer {
-  final String name;
-  final String description;
+  final String? name;
+  final String? description;
 
-  final List<ElementLayer> children = [];
+  final List<ElementLayer?> children = [];
 
-  ElementLayer({@required this.name, this.description = ''});
+  ElementLayer({required this.name, this.description = ''});
 
-  static ElementLayer fromJson(Map<String, dynamic> json) =>
-      LayerType.values.firstWhere((element) => element.toString() == json['type']).fromJson(json);
+  static ElementLayer? fromJson(Map<String, dynamic>? json) =>
+      LayerType.values.firstWhere((element) => element.toString() == json!['type']).fromJson(json);
   LayerType get type;
 }
 
 enum LayerType { container, text, paint }
 
-extension LayerTypeExtension on LayerType {
+extension LayerTypeExtension on LayerType? {
   String get name {
     switch (this) {
       case LayerType.container:
@@ -29,25 +29,25 @@ extension LayerTypeExtension on LayerType {
         return "Text";
       case LayerType.paint:
         return "Paint";
+      default:
+        return "";
     }
-    return "";
   }
 
-  IconData get icon {
+  IconData? get icon {
     switch (this) {
       case LayerType.container:
         return Mdi.group;
       case LayerType.text:
         return Mdi.text;
-        break;
       case LayerType.paint:
         return Mdi.paletteOutline;
-        break;
+      default:
+        return null;
     }
-    return null;
   }
 
-  ElementLayer create(String name, String description) {
+  ElementLayer? create(String name, String description) {
     switch (this) {
       case LayerType.container:
         return LayerContainer(name: name, description: description);
@@ -55,20 +55,22 @@ extension LayerTypeExtension on LayerType {
         return GroupElement(name: name, description: description);
       case LayerType.paint:
         return PaintElement(name: name, description: description);
+      default:
+        return null;
     }
-    return null;
   }
 
-  ElementLayer fromJson(Map<String, dynamic> json) {
+  ElementLayer? fromJson(Map<String, dynamic>? json) {
     switch (this) {
       case LayerType.container:
-        return LayerContainer.fromJson(json);
+        return LayerContainer.fromJson(json!);
       case LayerType.text:
-        return GroupElement.fromJson(json);
+        return GroupElement.fromJson(json!);
       case LayerType.paint:
-        return PaintElement.fromJson(json);
+        return PaintElement.fromJson(json!);
+      default:
+        return null;
     }
-    return null;
   }
 
   Map<String, dynamic> toJson(Map<String, dynamic> json) {
