@@ -2,6 +2,6 @@ import 'package:butterfly/models/packs/element.dart';
 import 'package:flutter/material.dart';
 
 class BackgroundPack extends ElementPack {
-  Image image;
-  String data;
+  Image? image;
+  String? data;
 }
