@@ -2,5 +2,5 @@ import 'package:butterfly/models/elements/type.dart';
 import 'package:butterfly/models/packs/element.dart';
 
 class LayerPack extends ElementPack {
-  ElementLayer element;
+  ElementLayer? element;
 }
