@@ -4,9 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class CreateLayerDialog extends StatefulWidget {
-  final ElementLayer parent;
+  final ElementLayer? parent;
 
-  CreateLayerDialog({Key key, this.parent}) : super(key: key);
+  CreateLayerDialog({Key? key, this.parent}) : super(key: key);
   @override
   _CreateLayerDialogState createState() => _CreateLayerDialogState();
 }
@@ -14,7 +14,7 @@ class CreateLayerDialog extends StatefulWidget {
 class _CreateLayerDialogState extends State<CreateLayerDialog> {
   final TextEditingController _nameController = TextEditingController();
   final TextEditingController _descriptionController = TextEditingController();
-  LayerType _type = LayerType.container;
+  LayerType? _type = LayerType.container;
 
   @override
   Widget build(BuildContext context) {
@@ -31,7 +31,7 @@ class _CreateLayerDialogState extends State<CreateLayerDialog> {
               onPressed: () {
                 BlocProvider.of<DocumentBloc>(context).add(LayerCreated(
                     parent: widget.parent,
-                    layer: _type.create(_nameController.text, _descriptionController.text)));
+                    layer: _type!.create(_nameController.text, _descriptionController.text)!));
                 Navigator.of(context).pop();
               },
               child: Text("CREATE"))
